Solution to Logical AND Operation

+++ Logical AND Operation@@67882a731b4392d8fc1577dc/code.c
@@ -3,8 +3,8 @@
 
 
 int main() {
-    int a,b;
-    scanf("%d %d",&a,&b);
+    double a,b;
+    scanf("%lf %lf",&a,&b);
     if(a && b>0){
         printf("True");
     }
